Conflicts:
	Sources/Internal/Entity/SceneSystem.h
	Sources/Internal/Math/Matrix4.h
	Sources/Internal/Render/Highlevel/RenderBatch.cpp
	Sources/Internal/Render/OGLHelpers.h
	Sources/Internal/Render/OcclusionQuery.cpp
	Sources/Internal/Render/RenderOptions.h

Former-commit-id: 36488085bf2e3b152542ca2f5c301d02d38dc013

+++ Sources/Internal/Platform/Process.cpp
@@ -171,7 +171,7 @@ namespace DAVA
 			String runArgsFlat = executablePath.GetAbsolutePathname();
 			if(runArgs.size() > 0)
 			{
-				for(int i = 0; i < runArgs.size(); ++i)
+				for(int i = 0; i < (int)runArgs.size(); ++i)
 				{
 					runArgsFlat += " ";
 					runArgsFlat += runArgs[i];
